requirements.txt
flask>=3.0
flask_sqlalchemy>=3.1
flask_wtf>=1.2
flask_migrate>=4.0
flask_login>=0.6
flask_limiter>=3.7
itsdangerous>=2.2
python-dotenv>=1.0
werkzeug>=3.0
google-generativeai>=0.7.2
pydantic>=2.6
